app/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
prometheus-client==0.19.0
gunicorn==21.2.0
eventlet==0.33.3 